Addressed to issue #88. Add ability to ban users for limited time and to view in the 'Show blocked' list time util any specific used was blocked.
Rename comment block css selectors
Rename comment block select options variable name

Fix css is block control select

Make blocking durations as the constant

Suppress jsx-a11y/no-onchange linter warning

Add ability to disable/enable comments (#55)

Fix auth panel unit test

Add curly braces to make code more readable

Set 'no-lonely-if' eslint option to zero

Revert "Merge branch 'japanese' of github.com:umputun/remark"

This reverts commit 3ac250a8802e9d1299d5a0bdd4a97cd393884a36, reversing
changes made to 328117fa840633c7edfd4ac0cb35ca09cdb7ac0a.

Fix navigation to comment issue (#82) on mobile.

ISSUE-73: UI support for DeleteMe request

Closes #73

* Add the API post request deleteMe
* Add a trigger button
* Add an utils/email module with the template and window opener
* Add a dropdown component
* Add a button component
* Hide the user-id text and a logout bout in the dropdown
* Add a noop empty function to pass a default handler callback to components

Admin page to approve deleteme request

Logout user on deleteme request

another flaky ttl block test (#172)

see #120

Fix crossdomain clickoutside listener

Input fix for Edge issue 169 (#171)

add deleteme flag to jwt token

it should prevent misuse of laked token to request user's data removal

extend jwt claim with flags

simplify user parsing with MustGetUserInfo

chnage flatlf to panic from MustGetUserInfo

add tests

Close DataStore on cleanup in tests (#173)

Windows fs can't remove testDb until file still used.
Therefore we need to close datastore on cleanup.

Revert "Fix navigation to comment issue (#82) on mobile."

This reverts commit bc5096b60b490bf3aac5dff2ca9524900161dae3.

Fix comment countdown time calculation when user and server time are differs (#74)

Fix time difference calculation between client and server

feature/ext-mongo (#177)

* externalize mongo wrapper

* remove mongo env from drone

* export mongo test for coverage report

Revert "Fix time difference calculation between client and server"

This reverts commit af82a58b72c8d19c1c236369abce8bfbe250f5b1.

Reapply: Fix time difference calculation between client and server

Save server and client time defference in every API call

+++ backend/app/rest/api/admin.go
@@ -108,6 +108,12 @@ func (a *admin) deleteMeRequestCtrl(w http.ResponseWriter, r *http.Request) {
 
 	log.Printf("[INFO] delete all user comments by request for %s, site %s", claims.User.ID, claims.SiteID)
 
+	// deleteme set by deleteMeCtrl, this check just to make sure we not trying to delete with leaked token
+	if !claims.Flags.DeleteMe {
+		rest.SendErrorJSON(w, r, http.StatusForbidden, errors.New("forbidden"), "can't use provided token")
+		return
+	}
+
 	if err := a.dataService.DeleteUser(claims.SiteID, claims.User.ID); err != nil {
 		rest.SendErrorJSON(w, r, http.StatusBadRequest, err, "can't delete user")
 		return
